+++ tiltft-frontend/public/index.html
@@ -25,10 +25,7 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <link href="https://fonts.googleapis.com/css?family=Gloria+Hallelujah&display=swap" rel="stylesheet">
-<!--    <link-->
-<!--            href="https://fonts.googleapis.com/css?family=Material+Icons&display=block"-->
-<!--            rel="stylesheet"-->
-<!--    />-->
+
     <title>React App</title>
   </head>
   <body>
