+++ tiltft-frontend/src/components/RegionSelect/styles.scss
@@ -1,0 +1,4 @@
+.region-select {
+  height: 50px;
+  line-height: 50px;
+}
