+++ tiltft-frontend/src/scenes/Home/styles.scss
@@ -1,0 +1,9 @@
+.summoner-name-bar {
+  width: 300px;
+  line-height: 50px;
+  height: 50px;
+}
+
+p, h2, .summoner-name-bar {
+  font-family: "Gloria Hallelujah", sans-serif;
+}
